// src/app/Protegido/interfaces/clienteInterface.ts
// Generated by https://quicktype.io

export interface ClienteI {
  ok:      boolean;
  cliente: Cliente[];
}

export interface ClienteShow{
  ok:string;
  msg?:string
  cliente?:Cliente
}

export interface Cliente {
  id:         number;
  nombre:     string;
  apellido_p: string;
  apellido_m: string;
  direccion:  string;
  genero:     string;
  createdAt?:  string;
  updatedAt?:  string;
  deletedAt?:  null;
}