// src/app/Protegido/interfaces/medicamentosI.ts
// Generated by https://quicktype.io





export interface MedicamentoSaveI {
  ok:          boolean;
  msg?:         string;
  medicamento?:Medicina
}


export interface MedicamentoInterface {
  ok:      boolean;
  medicina: Medicina[];// Generated by https://quicktype.io
}

export interface Medicina {
  id:                number | null;
  usuario_id:        number;
  tipo_id:           number;
  nombre_cientifico: string;
  nombre_registrado: string;
  fecha_caducidad:   string;
  codigo:            string;
  cantidad:          number;
  precio:            number;
  imagen:            string | null;
  simbolo:           null | string;
  createdAt:         string | null;
  updatedAt:         null | string;
  deletedAt:         null | string;
  Usuario?:           Usuario;
  Tipo_medicamento?:  TipoMedicamento | null;
}

export interface TipoMedicamento {
  id:          number;
  tipo:        string;
  descripcion: string;
  createdAt:   string;
  updatedAt:   string;
  deletedAt:   null;
}

export interface Usuario {
  id?:         number;
  nombre?:     string;
  email?:      string;
  edad?:       number;
  apellido_p?: string;
  apellido_m?: string;
  password?:   string;
  estado?:     boolean;
  createdAt:  string;
  updatedAt:  string;
  deletedAt:  null;
}


export interface modalDetalleDataI{
  id:number;
  tabla:string;
}